requirements.txt
# Core Agentic Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# LLM & API
anthropic>=0.39.0

# Data Ingestion
yt-dlp>=2024.8.0
feedparser>=6.0.11
newspaper3k>=0.2.8
beautifulsoup4>=4.12.3
lxml>=5.3.0
requests>=2.32.0

# Audio Processing & Transcription
# Using faster-whisper instead of openai-whisper (more stable on Mac, still local)
faster-whisper>=1.0.0
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Embeddings & Vector Search
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
torch>=2.0.0

# Data Processing
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.2

# Testing & Development (optional)
pytest>=8.0.0
black>=24.0.0
